model/agent/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
google-generativeai
python-dotenv
pillow
requests
